Refactor: X아이콘 삭제, o-auth디렉토리 분리, oauthUrl 타입 수정, loginconfig삭제, OAuthoButton태그 수정,
