+++ app/login/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import OAuthButton from './_components/o-auth-button';
+import OAuthButton from './_components/oauth-button/o-auth-button';
 
 const Login = () => {
   return (
@@ -8,8 +8,8 @@ const Login = () => {
       <div className="flex-column h-96 w-96 min-w-96 space-y-8 p-1 ">
         <h1 className="text-center text-5xl">동구라미 로고</h1>
         <div className="rounded-md border border-slate-200 p-8">
-          <OAuthButton domain="kakao"></OAuthButton>
-          <OAuthButton domain="naver"></OAuthButton>
+          <OAuthButton domain="kakao" />
+          <OAuthButton domain="naver" />
         </div>
       </div>
     </div>
